testing/web-platform/tests/video-rvfc/request-video-frame-callback-parallel.html

<!DOCTYPE html>
<html>
<title>Test having multiple video.rVFC callbacks in flight for a single element.</title>
<body></body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/media.js"></script>
<script>

promise_test(async function(t) {
    let done;
    const promise = new Promise(resolve => done = resolve);

    let video = document.createElement('video');
    document.body.appendChild(video);

    let firstTime;
    let firstMetadata;

    video.requestVideoFrameCallback(t.step_func((time, metadata) => {
      firstTime = time;
      firstMetadata = metadata;
    }));

    video.requestVideoFrameCallback(t.step_func((time, metadata) => {
      assert_equals(firstTime, time);
      assert_equals(firstMetadata.presentationTime, metadata.presentationTime);
      assert_equals(JSON.stringify(firstMetadata), JSON.stringify(metadata));
      assert_object_equals(firstMetadata, metadata);
      done();
    }));

    video.src = getVideoURI('/media/movie_5');
    video.play();

    return promise;
}, 'Test callbacks get the same information.');

</script>
</html>
